re-add init function for home manager
my stupid ass in b979657 thought that
home-manager.nixosModules.home-manager was a function that returned a
module instead of just a module (which is also a function but that's
besides the point). i forgot to consider that the home-manager module
adds a custom home-manager option set, nor did i realize that a flake
module could be a plain attribute set.