home/base: enable Git maintenance
This is enabled in home-manager becuase Git attempts to hardcode the
path when it is done through Git. Enabling Git in home-manager creates a
managed config at $XDG_CONFIG_HOME/git/config, but if ~/.gitconfig
exists, git will use that instead which still allows for mutable config.
The activation script is configured to automatically create ~/.gitconfig
if it doesn't exist, so everything should still work exactly as before.