home(neovim): add gcc to env
A compiler is required for the compilation of treesitter grammars. I
should probably figure out how to just add every grammar to Neovim, but
this works with my existing configuration for now. I'm choosing to keep
my config separate because I want to still be able to use it on systems
without Nix.